Moved hashing the received message to cothorityd, now HashHash makes sense again (cothorityd hashes the received message, which again is hash by default in the cli; @nikkolasg: don't worry it is still would be possible to send the original message, we could add a flag to the CLI for that later)
